# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Basic WebRTC API
  description: REST API for the Basic WebRTC video conferencing application.
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /api/rooms:
    get:
      tags: [Rooms]
      summary: List rooms the caller hosts or participated in (dashboard "Recent meetings")
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Up to 10 most recent rooms with live participant counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        hostId:
                          type: string
                        title:
                          type: string
                        isLocked:
                          type: boolean
                        maxParticipants:
                          type: integer
                        participantCount:
                          type: integer
                        hostName:
                          type: string
                          nullable: true
                        createdAt:
                          type: string
                          format: date-time
                        endedAt:
                          type: string
                          format: date-time
                          nullable: true
    post:
      tags: [Rooms]
      summary: Create a new room
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                type:
                  type: string
                  enum: [open, locked, waiting-room]
      responses:
        "201":
          description: Room created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RoomTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Room-scoped JWT (the same token used for the WebSocket upgrade), not an
        account access token. Used by the in-call caption upload.
...